[gvar] minor

+++ src/hb-ot-var-gvar-table.hh
@@ -655,8 +655,8 @@ no_more_gaps:
 	/* apply specified / inferred deltas to points */
 	for (unsigned int i = 0; i < points.length; i++)
 	{
-	  points[i].x += (float) roundf (deltas[i].x);
-	  points[i].y += (float) roundf (deltas[i].y);
+	  points[i].x += roundf (deltas[i].x);
+	  points[i].y += roundf (deltas[i].y);
 	}
       } while (iterator.move_to_next ());
